requirements.txt
validators==0.20.0
python-telegram-bot==13.15
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.10.0
webdriver-manager==4.0.1
selenium-stealth==1.0.6
urllib3==1.26.16
